+++ docs/dialect/graphblas_dialect_tutorials/graphblas_lower/index.rst
@@ -5,14 +5,15 @@ GraphBLAS Lowering Pass
 
 These tutorials and examples will cover how to use the GraphBLAS dialect's ``--graphblas-lower`` pass by using the :ref:`engine` to lower several ops from the GraphBLAS dialect.
 
-The content of the tutorials are somewhat sequentially dependent as some later tutorials assume completion of previous tutorials.
+The main purpose of the ``--graphblas-lower`` pass is to lower from the GraphBLAS dialect into a lower level dialect, e.g. the `SCF dialect <https://mlir.llvm.org/docs/Dialects/SCFDialect/>`_ or the `Sparse Tensor dialect <https://mlir.llvm.org/docs/Dialects/SparseTensorOps/>`_. Rather than simply showing the code transformations, we'll use the :ref:`engine` to take some example MLIR code using the GraphBLAS dialect and create executable Python code from it. Since we won't be able to go over all of the ops in the GraphBLAS dialect and since all of the ops are documented with examples in the :ref:`graphblas_ops_reference`, our examples in this section will mostly cover ops that are commonly used.
+
+The content of the tutorials are somewhat sequentially dependent as some later tutorials assume completion of previous tutorials. They also assume familiarity with the content of the :ref:`graphblas_ops_reference`.
 
 .. toctree::
    :maxdepth: 1
 
-   graphblas_convert_layout
-   graphblas_matrix_select
-   graphblas_matrix_reduce_to_scalar
-   graphblas_apply
-   graphblas_matrix_multiply
-   graphblas_matrix_multiply_reduce_to_scalar
+   python_utilities
+   sparse_layouts
+   vector_ops
+   matrix_ops
+   debugging_ops
