+++ mlir_graphblas/tools/__init__.py
@@ -1,1 +1,2 @@
 from .tersify_mlir import *
+from . import utils
